01-tutorials/03-deployment/03-agentcore-deployment/pyproject.toml
[project]
name = "03-agentcore-deployment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore>=0.1.2",
    "bedrock-agentcore-starter-toolkit>=0.1.6",
    "python-dotenv>=1.1.1",
    "strands-agents>=1.5.0",
]
